+++ rocAL/rocAL/include/meta_data/bounding_box_graph.h
@@ -30,6 +30,6 @@ public:
     void process(MetaDataBatch* meta_data, bool segmentation) override;
     void update_meta_data(MetaDataBatch* meta_data, decoded_image_info decode_image_info, bool segmentation) override;
     void update_random_bbox_meta_data(MetaDataBatch* meta_data, decoded_image_info decoded_image_info,crop_image_info crop_image_info) override;
-    void update_box_encoder_meta_data(std::vector<float> *anchors, pMetaDataBatch full_batch_meta_data ,float criteria, bool offset , float scale, std::vector<float>& means, std::vector<float>& stds) override;
+    void update_box_encoder_meta_data(std::vector<double> *anchors, pMetaDataBatch full_batch_meta_data ,double criteria, bool offset , double scale, std::vector<double>& means, std::vector<double>& stds) override;
 };
